requirements.txt
faster-whisper==1.0.3
gradio==4.42.0
edge-tts==6.1.12
openai==1.43.0
PyYAML==6.0.2